requirements.txt
openai>=1.0.0
qiskit>=0.45.0
qiskit-ibm-runtime>=0.15.0
qiskit-aer>=0.12.0
numpy>=1.24.0
mcp>=0.4.0
pydantic>=2.0.0
python-dotenv>=0.19.0